requirements.txt
Django~=3.2.10
djangorestframework~=3.13.1
djongo~=1.3.6
django-filter~=21.1
pymongo~=3.12.3
numpy~=1.22.3
pytesseract~=0.3.9
Pillow~=9.1.0
djoser~=2.1.0
six~=1.16.0
websocket-client~=0.59.0
beautifulsoup4~=4.11.1
soupsieve~=2.3.2.post1
dnspython~=2.2.1
idna~=3.3
requests~=2.27.1
cryptography~=36.0.2
matplotlib~=3.5.2
scikit-image~=0.19.2
defusedxml~=0.7.1
cffi~=1.15.0
pip~=21.1.2
wheel~=0.36.2
rsa~=4.8
pyasn1~=0.4.8
django-mptt~=0.13.4
pytz~=2022.1
scipy~=1.8.1
tika~=1.24
Wand~=0.6.7
PyYAML~=5.4.1
attrs~=21.4.0
Faker~=13.12.0
python-dateutil~=2.8.2
setuptools~=57.0.0
torch~=1.11.0
django-model-utils~=4.2.0
torchvision~=0.12.0
urllib3~=1.26.9
networkx~=2.8.2
distro~=1.7.0
sqlparse~=0.2.4
asgiref~=3.5.0
Jinja2~=3.1.1
docker~=5.0.3
paramiko~=2.11.0
flake8~=4.0.1
pyflakes~=2.4.0
cachetools~=5.0.0
certifi~=2021.10.8
google-api-python-client~=2.47.0
MarkupSafe~=2.1.1
martor~=1.6.13
Markdown~=3.3.4
httplib2~=0.20.4
pandas~=1.4.2
PyPDF2~=1.28.2
django-allauth~=0.50.0
PyJWT~=2.3.0
chardet~=4.0.0
docker-compose~=1.29.2
dockerpty~=0.4.1
texttable~=1.6.4
docopt~=0.6.2
jsonschema~=3.2.0
python-dotenv~=0.20.0
itypes~=1.2.0
coreapi~=2.3.3
coreschema~=0.0.4
uritemplate~=4.1.1
easyocr~=1.4.2
tifffile~=2022.5.4
imageio~=2.19.3
packaging~=21.3
djangorestframework-recursive~=0.1.2
inflection~=0.5.1
gunicorn~=20.1.0
pyparsing~=3.0.8
jwcrypto~=1.2
Deprecated~=1.2.13
oauthlib~=3.2.0
bcrypt~=3.2.2
fonttools~=4.33.3
pycparser~=2.21
django-rest-auth~=0.9.5
wrapt~=1.14.1
fuzzywuzzy~=0.18.0
pyrsistent~=0.18.1
cycler~=0.11.0
kiwisolver~=1.4.2
mongoengine~=0.24.1
djangotoolbox~=1.8.0
jsonfield~=3.1.0
django-notifications-hq~=1.7.0
swapper~=1.3.0
django-oauth-toolkit~=2.0.0
environ==1.0
facebook-sdk==3.1.0
